Fix overflow bug and tidy up main function

+++ player.c
@@ -5,7 +5,7 @@
 static const SongEvent* current_song;
 static uint16_t current_state[CHANNELS][3] = {}; // [channel index] = {current_sample, counter, target}
 static uint16_t current_sample = 0;
-static uint16_t current_sample_counter = 0;
+static uint32_t current_sample_counter = 0;
 static uint16_t current_pos = 0;
 
 void set_song(const SongEvent song[]) {
